+++ saleor/graphql/product/types/products.py
@@ -430,6 +430,12 @@ class ProductVariant(ChannelContextTypeWithMetadata, CountableDjangoObjectType):
 
 @key(fields="id")
 class Product(ChannelContextTypeWithMetadata, CountableDjangoObjectType):
+    description_json = graphene.JSONString(
+        description="Description of the product (JSON).",
+        deprecation_reason=(
+            "Will be removed in Saleor 4.0. Use the `description` field instead."
+        ),
+    )
     url = graphene.String(
         description="The storefront URL for the product.",
         required=True,
@@ -497,7 +503,6 @@ class Product(ChannelContextTypeWithMetadata, CountableDjangoObjectType):
             "category",
             "charge_taxes",
             "description",
-            "description_json",
             "id",
             "name",
             "slug",
@@ -531,6 +536,10 @@ class Product(ChannelContextTypeWithMetadata, CountableDjangoObjectType):
         if category_id is None:
             return None
         return CategoryByIdLoader(info.context).load(category_id)
+
+    @staticmethod
+    def resolve_description_json(root: ChannelContext[models.Product], info):
+        return root.node.description
 
     @staticmethod
     def resolve_tax_type(root: ChannelContext[models.Product], info):
@@ -881,6 +890,12 @@ class ProductType(CountableDjangoObjectType):
 
 @key(fields="id")
 class Collection(ChannelContextTypeWithMetadata, CountableDjangoObjectType):
+    description_json = graphene.JSONString(
+        description="Description of the collection (JSON).",
+        deprecation_reason=(
+            "Will be removed in Saleor 4.0. Use the `description` field instead."
+        ),
+    )
     products = ChannelContextFilterConnectionField(
         Product,
         filter=ProductFilterInput(description="Filtering options for products."),
@@ -905,7 +920,6 @@ class Collection(ChannelContextTypeWithMetadata, CountableDjangoObjectType):
         description = "Represents a collection of products."
         only_fields = [
             "description",
-            "description_json",
             "id",
             "name",
             "seo_description",
@@ -946,9 +960,19 @@ class Collection(ChannelContextTypeWithMetadata, CountableDjangoObjectType):
     def __resolve_reference(root, _info, **_kwargs):
         return graphene.Node.get_node_from_global_id(_info, root.id)
 
+    @staticmethod
+    def resolve_description_json(root: ChannelContext[models.Collection], info):
+        return root.node.description
+
 
 @key(fields="id")
 class Category(CountableDjangoObjectType):
+    description_json = graphene.JSONString(
+        description="Description of the category (JSON).",
+        deprecation_reason=(
+            "Will be removed in Saleor 4.0. Use the `description` field instead."
+        ),
+    )
     ancestors = PrefetchingConnectionField(
         lambda: Category, description="List of ancestors of the category."
     )
@@ -979,7 +1003,6 @@ class Category(CountableDjangoObjectType):
         )
         only_fields = [
             "description",
-            "description_json",
             "id",
             "level",
             "name",
@@ -994,6 +1017,10 @@ class Category(CountableDjangoObjectType):
     @staticmethod
     def resolve_ancestors(root: models.Category, info, **_kwargs):
         return root.get_ancestors()
+
+    @staticmethod
+    def resolve_description_json(root: models.Category, info):
+        return root.description
 
     @staticmethod
     def resolve_background_image(root: models.Category, info, size=None, **_kwargs):