+++ saleor/page/models.py
@@ -3,7 +3,7 @@ from django.db import models
 from ..core.db.fields import SanitizedJSONField
 from ..core.models import ModelWithMetadata, PublishableModel
 from ..core.permissions import PagePermissions, PageTypePermissions
-from ..core.sanitizers.editorjs_sanitizer import clean_editor_js
+from ..core.utils.editorjs import clean_editor_js
 from ..core.utils.translations import TranslationProxy
 from ..seo.models import SeoModel, SeoModelTranslation
 
@@ -14,10 +14,7 @@ class Page(ModelWithMetadata, SeoModel, PublishableModel):
     page_type = models.ForeignKey(
         "PageType", related_name="pages", on_delete=models.CASCADE
     )
-    content = models.TextField(blank=True)
-    content_json = SanitizedJSONField(
-        blank=True, default=dict, sanitizer=clean_editor_js
-    )
+    content = SanitizedJSONField(blank=True, default=dict, sanitizer=clean_editor_js)
     created = models.DateTimeField(auto_now_add=True)
 
     translated = TranslationProxy()
@@ -36,10 +33,7 @@ class PageTranslation(SeoModelTranslation):
         Page, related_name="translations", on_delete=models.CASCADE
     )
     title = models.CharField(max_length=255, blank=True)
-    content = models.TextField(blank=True)
-    content_json = SanitizedJSONField(
-        blank=True, default=dict, sanitizer=clean_editor_js
-    )
+    content = SanitizedJSONField(blank=True, default=dict, sanitizer=clean_editor_js)
 
     class Meta:
         ordering = ("language_code", "page", "pk")