+++ CHANGELOG.md
@@ -34,9 +34,11 @@ All notable, unreleased changes to this project will be documented in this file.
 - Get tax rate from plugins - #6649 by @IKarbowiak
 - Added support for querying user by email - #6632 @LeOndaz
 - Add order shipping tax rate - #6678 by @IKarbowiak
+- Deprecate field `descriptionJSON` from `Product`, `Category`, `Collection` and field `contentJSON` from `Page` - #6692 by @d-wysocki
 - Fix products visibility - #6704 by @IKarbowiak
 - Introduce page reference attributes - #6624 by @IKarbowiak
 - Introduce product reference attributes - #6711 by @IKarbowiak
+- Add metadata to warehouse - #6727 by @d-wysocki
 
 # 2.11.1